% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims.R
\name{binGlmP}
\alias{binGlmP}
\title{Model simulation for binomial generalised linear model with two level categorical predictor}
\usage{
binGlmP(n, f)
}
\arguments{
\item{n}{sample size}

\item{f}{effect size (probability)}
}
\value{
p value for model simulation
}
\description{
Model simulation for binomial generalised linear model with two level categorical predictor
}
