% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims.R
\name{binGlmPRand}
\alias{binGlmPRand}
\title{Model simulation for binomial generalised linear model with two level categorical predictor and a random intercept fitted}
\usage{
binGlmPRand(n, f, ind)
}
\arguments{
\item{n}{total number of trials}

\item{f}{effect size}

\item{ind}{number of participants}
}
\value{
p value for model simulation
}
\description{
Model simulation for binomial generalised linear model with two level categorical predictor and a random intercept fitted
}
