% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims.R
\name{expBin2}
\alias{expBin2}
\title{Simulate experimental data with two conditions and a binomial outcome}
\usage{
expBin2(n, prob1, prob2)
}
\arguments{
\item{n}{sample size}

\item{prob1}{Probability of success for control group}

\item{prob2}{Probability of success for experimental group}
}
\description{
Simulate experimental data with two conditions and a binomial outcome
}
