% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims.R
\name{expBin2Rand}
\alias{expBin2Rand}
\title{Simulate experimental data with two conditions and a binomial outcome}
\usage{
expBin2Rand(n, prob1, prob2, ind)
}
\arguments{
\item{n}{sample size}

\item{prob1}{Probability of success for control group}

\item{prob2}{Probability of success for experimental group}

\item{ind}{number of participants}
}
\description{
Simulate experimental data with two conditions and a binomial outcome
}
