% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims.R
\name{expBin3Rand}
\alias{expBin3Rand}
\title{Simulate experimental data with three conditions and a binomial outcome}
\usage{
expBin3Rand(n, prob1, prob2, prob3, ind)
}
\arguments{
\item{n}{sample size}

\item{prob1}{Probability of success for group 1}

\item{prob2}{Probability of success for group 2}

\item{prob3}{Probability of success for group 3}

\item{ind}{number of participants}
}
\description{
Simulate experimental data with three conditions and a binomial outcome
}
