% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims.R
\name{expCont2}
\alias{expCont2}
\title{Simulate experimental data with two conditions and a continuous outcome}
\usage{
expCont2(n, mean1, var1, mean2, var2)
}
\arguments{
\item{n}{sample size}

\item{mean1}{mean of dependent variable for control group}

\item{var1}{standard deviation of dependent variable for experimental group}

\item{mean2}{mean of dependent variable for experomental group}

\item{var2}{standard deviation of dependent variable for experimental group}
}
\description{
Simulate experimental data with two conditions and a continuous outcome
}
