% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims.R
\name{expCont2Rand}
\alias{expCont2Rand}
\title{Simulate experimental data with two conditions, a continuous outcome varible and a participant index}
\usage{
expCont2Rand(n, mean1, var1, mean2, var2, ind)
}
\arguments{
\item{n}{sample size}

\item{mean1}{mean of dependent variable for control group}

\item{var1}{standard deviation of dependent variable for control group}

\item{mean2}{mean of dependent variable for experimental group}

\item{var2}{standard deviation of dependent variable for experimental group}

\item{ind}{number of participants}
}
\description{
Simulate experimental data with two conditions, a continuous outcome varible and a participant index
}
