% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims.R
\name{expCont3}
\alias{expCont3}
\title{Simulate experimental data with three conditions and a continuous outcome}
\usage{
expCont3(n, mean1, var1, mean2, var2, mean3, var3)
}
\arguments{
\item{n}{sample size}

\item{mean1}{mean of dependent variable for experimental group 1}

\item{var1}{standard deviation for experimental group2}

\item{mean2}{mean of dependent variable for experimental group 2}

\item{var2}{standard deviation for experimental group 2}

\item{mean3}{mean of dependent variable for experimental group 3}

\item{var3}{standard deviation for experimental group 3}
}
\description{
Simulate experimental data with three conditions and a continuous outcome
}
