% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims.R
\name{expCont3Rand}
\alias{expCont3Rand}
\title{Simulate experimental data with three conditions, a continuous outcome and a participant index}
\usage{
expCont3Rand(n, mean1, var1, mean2, var2, mean3, var3, ind)
}
\arguments{
\item{n}{sample size}

\item{mean1}{mean of dependent variable for experimental group1}

\item{var1}{standard deviation of dependent variable for experimental group1}

\item{mean2}{mean of dependent variable for experimental group2}

\item{var2}{standard deviation of dependent variable for experimental group2}

\item{mean3}{mean of dependent variable for experimental group3}

\item{var3}{standard deviation of dependent variable for experimental group 3}

\item{ind}{number of participants}
}
\description{
Simulate experimental data with three conditions, a continuous outcome and a participant index
}
