% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims.R
\name{glmME}
\alias{glmME}
\title{Simulation of margin of error for binomial generalised linear model}
\usage{
glmME(n, f)
}
\arguments{
\item{n}{sample size}

\item{f}{effect size}
}
\value{
margin of error
}
\description{
Simulation of margin of error for binomial generalised linear model
}
