% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims.R
\name{lmME}
\alias{lmME}
\title{Simulation of confidence intervals for basic linear model}
\usage{
lmME(n, f)
}
\arguments{
\item{n}{sample size}

\item{f}{effect size}
}
\value{
confidence interval
}
\description{
Simulation of confidence intervals for basic linear model
}
