% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims.R
\name{lmP}
\alias{lmP}
\title{Simulation for a linear model that returns only p values}
\usage{
lmP(n, f)
}
\arguments{
\item{n}{sample size}

\item{f}{effect size}
}
\value{
p-value for effect
}
\description{
Simulation for a linear model that returns only p values
}
