% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims.R
\name{lmSim}
\alias{lmSim}
\title{Simulation for a linear model
Arguments: n=sample size, and f=effect size between a control and experiment
group}
\usage{
lmSim(n, f)
}
\description{
Simulation for a linear model
Arguments: n=sample size, and f=effect size between a control and experiment
group
}
