% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims.R
\name{metaRsim}
\alias{metaRsim}
\title{Model simulation for meta analysis random effects model}
\usage{
metaRsim(n, f, var)
}
\arguments{
\item{n}{sample size}

\item{f}{effect size}
}
\value{
p value for model simulation
}
\description{
Model simulation for meta analysis random effects model
}
