% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims.R
\name{power2}
\alias{power2}
\title{Simulation for binomial generalised linear model
Simulation for binomial generalised linear model that only returns p values
Simulation of poisson generalised linear model
Simulation os Poissons generalised linear model that only returns p values
Simulation of lm with random effects variable
Simulation of lm with random effects returning only p values
Simulation of binomial glm with random effects variable
Simulation of binomial glm with random effects variable returning only p values
Simulation of poisson glm with random effects
Simulation of poisson glm with random effects returning only p values
Power simulation from model simulation}
\usage{
power2(mod)
}
\arguments{
\item{Model}{simulation}
}
\value{
Statistical power value for model
}
\description{
Simulation for binomial generalised linear model
Simulation for binomial generalised linear model that only returns p values
Simulation of poisson generalised linear model
Simulation os Poissons generalised linear model that only returns p values
Simulation of lm with random effects variable
Simulation of lm with random effects returning only p values
Simulation of binomial glm with random effects variable
Simulation of binomial glm with random effects variable returning only p values
Simulation of poisson glm with random effects
Simulation of poisson glm with random effects returning only p values
Power simulation from model simulation
}
