% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims.R
\name{powerBinGlm}
\alias{powerBinGlm}
\title{Power simulation for binomial generalised linear model with two level categorical predictor}
\usage{
powerBinGlm(n, f, iter)
}
\arguments{
\item{n}{sample size}

\item{f}{effect size}

\item{iter}{number of iterations of simulation}
}
\value{
statistical power value for model
}
\description{
Power simulation for binomial generalised linear model with two level categorical predictor
}
