% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims.R
\name{powerBinGlmRand}
\alias{powerBinGlmRand}
\title{Power analysis for binomial generalised linear model with two level categorical predictor and random intercept}
\usage{
powerBinGlmRand(n, f, ind, iter = 100)
}
\arguments{
\item{n}{total number of trials}

\item{f}{effect size}

\item{ind}{number of participants}

\item{iter}{number of iterations of model simulation}
}
\value{
statistical power value for model
}
\description{
Power analysis for binomial generalised linear model with two level categorical predictor and random intercept
}
