% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims.R
\name{powerMetaR}
\alias{powerMetaR}
\title{Power simulation for random effects meta analysis model}
\usage{
powerMetaR(n, f, var, iter = 100)
}
\arguments{
\item{n}{sample size}

\item{f}{effect size}

\item{var}{variance for meta analysis model}

\item{iter}{number of iterations of simulation}
}
\value{
statistical power for meta analysis
}
\description{
Power simulation for random effects meta analysis model
}
