% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims.R
\name{powGlmPlot}
\alias{powGlmPlot}
\title{Plot distribution of p-values from binomial generalised linear model}
\usage{
powGlmPlot(n, f, iter = 100)
}
\arguments{
\item{n}{sample size}

\item{f}{effect size}

\item{iter}{number of iterations of simulation}
}
\value{
plot of distribution of p-values
}
\description{
Plot distribution of p-values from binomial generalised linear model
}
