% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims.R
\name{powGlmRandPlot}
\alias{powGlmRandPlot}
\title{Plot p-values for generalised binomial linear model with a single binomial predictor and a random intercept}
\usage{
powGlmRandPlot(n, f, ind, iter = 100)
}
\arguments{
\item{n}{total number of trials}

\item{f}{effect size}

\item{ind}{number of participants}

\item{iter}{number of iterations of simulation}
}
\value{
distribution plot of p-values
}
\description{
Plot p-values for generalised binomial linear model with a single binomial predictor and a random intercept
}
