% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims.R
\name{powLmPlot}
\alias{powLmPlot}
\title{Plot p-values from basic linear model simulations}
\usage{
powLmPlot(n, f, iter)
}
\arguments{
\item{n}{sample size}

\item{f}{effect size}

\item{iter}{number of iterations of simulation}
}
\value{
plot of p values
}
\description{
Plot p-values from basic linear model simulations
}
