% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims.R
\name{powMetaPlot}
\alias{powMetaPlot}
\title{Plot distribution of p values from meta analysis random effects model}
\usage{
powMetaPlot(n, f, var, iter = 100)
}
\arguments{
\item{n}{sample size}

\item{f}{effect size}

\item{var}{variance for meta analysis model}

\item{iter}{number of iterations of simulation}
}
\value{
plot of distribution of p values
}
\description{
Plot distribution of p values from meta analysis random effects model
}
