% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims.R
\name{powPlot}
\alias{powPlot}
\title{Plot p-values from basic linear model simulations}
\usage{
powPlot(n, f)
}
\arguments{
\item{n}{sample size}

\item{f}{effect size}
}
\value{
plot of p values
}
\description{
Plot p-values from basic linear model simulations
}
