% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims.R
\name{precisionglm}
\alias{precisionglm}
\title{Precision simulation for binomial generalised linear model with a two level categorical predictor}
\usage{
precisionglm(n, f, iter = 100)
}
\arguments{
\item{n}{sample size}

\item{f}{effect size}

\item{iter}{number of iterations of simulation}
}
\value{
Statistical precision of model
}
\description{
Precision simulation for binomial generalised linear model with a two level categorical predictor
}
