% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims.R
\name{precisionlm}
\alias{precisionlm}
\title{Precision simulation for basic linear model with a two level categorical predictor}
\usage{
precisionlm(n, f, iter = 100)
}
\arguments{
\item{n}{sample size}

\item{f}{effect size}

\item{iter}{number of iterations of simulation}
}
\value{
Statistical precision of a model
}
\description{
Precision simulation for basic linear model with a two level categorical predictor
}
