% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims.R
\name{precisionlmPlot}
\alias{precisionlmPlot}
\title{Plot the distribution of margin of errors from a linear model simulation}
\usage{
precisionlmPlot(n, f, iter = 100)
}
\arguments{
\item{n}{sample size}

\item{f}{effect size}

\item{iter}{number of iterations of simulation}
}
\value{
plot of distribution of margin of errors
}
\description{
Plot the distribution of margin of errors from a linear model simulation
}
