% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims.R
\name{pSim}
\alias{pSim}
\title{Average p value simulation for a model function}
\usage{
pSim(mod)
}
\arguments{
\item{mod}{A model simulation}
}
\value{
P-values across simulations
}
\description{
Average p value simulation for a model function
}
