% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims.R
\name{lmSim}
\alias{lmSim}
\title{The powR R package is currently still under development.
Functions for different model simulations are being created.}
\usage{
lmSim(n, f)
}
\arguments{
\item{n}{sample size}

\item{f}{effect size}
}
\value{
A summary table of the model results
}
\description{
Simulation for a linear model
}
