% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims.R
\name{power2}
\alias{power2}
\title{Power simulation from model simulation}
\usage{
power2(mod)
}
\arguments{
\item{Model}{simulation}
}
\value{
Statistical power value for model
}
\description{
Power simulation from model simulation
}
