% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims.R
\name{powerlm}
\alias{powerlm}
\title{Power simulation for basic linear model with a two level categorical predictor}
\usage{
powerlm(n, f, iter = 100)
}
\arguments{
\item{n}{sample size}

\item{f}{effect size}

\item{iter}{number of iterations of simulation}
}
\value{
Statistical power value for model
}
\description{
Power simulation for basic linear model with a two level categorical predictor
}
